+++ src/components/CortexPage/CortexPage.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import React from 'react';
-import { isNil } from 'lodash';
+import { capitalize, isNil } from 'lodash';
 import { CortexLayout } from '../CortexLayout';
 import { ScorecardsPage } from '../../extensions';
 import { SettingsPage } from '../SettingsPage';
@@ -29,6 +29,7 @@ import { extensionApiRef } from '../../api/ExtensionApi';
 import { HelpPage } from '../HelpPage';
 import { isBeforeShutdownDate, shouldShowExpirationBanner } from '../Entitlements/ExpirationUtils';
 import { ExpirationBanner } from '../Entitlements/ExpirationBanner';
+import { useInitiativesCustomName } from '../../utils/hooks';
 
 export const CortexPage = ({
   title = 'Cortex',
@@ -63,6 +64,8 @@ export const CortexPage = ({
     return await cortexApi.getExpiration();
   }, []);
 
+  const { plural: initiativesName } = useInitiativesCustomName();
+
   if (loadingPermissions || loadingHelpPage || loadingExpiration || isNil(expiration)) {
     return <Progress />;
   }
@@ -79,7 +82,7 @@ export const CortexPage = ({
         <CortexLayout.Route path="reports" title="Reports">
           <ReportsPage />
         </CortexLayout.Route>
-        <CortexLayout.Route path="initiatives" title="Initiatives">
+        <CortexLayout.Route path="initiatives" title={capitalize(initiativesName)}>
           <InitiativesPage />
         </CortexLayout.Route>
         {/*