+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Cortex page title and subtitle are now customizable
 - Adds `hideCortexLinks` flag to hide all links to Cortex app
+- Adds `initiativeNameOverride` config to override "Initiative" name
 
 ### 2.6.6